First-Commit
This project is how to install Ignite in a project using Postgres as and Ignite both. Postgres is usually used as DB so I have install that one and Ignite is also added to test and use it on 100,000 products and checking it's response time. Ignite is basically used here using a credentials.